services/gpg-agent: Disable SSH integration
A regression introduced in e18d7b1 (flake inputs bump, including
nixpkgs), caused `ssh-add` to fail when trying to add keys to the agent.
Disabling SSH integration in the gpg-agent resolves the issue.

I'm not quite sure if I even need the SSH integration, or even the GPG
agent altogether, but it can stay as long as I can add SSH keys to the
agent.